requirements.docker.txt
# ===== Final Requirements for Emergent Machine Pedagogy (v3 - Patched) =====
# Minimal set of top-level dependencies with a newer llama-cpp-python

# Core LLM and embedding libraries
llama-cpp-python==0.2.77  # <-- UPDATED VERSION
torch==2.1.2
sentence-transformers==2.2.2

# PINNED SUB-DEPENDENCIES to resolve ImportError
huggingface-hub<0.20.0
transformers<4.31.0

# Data handling and scientific computing
pandas==2.1.4
scikit-learn==1.3.2

# Utility for progress bars
tqdm==4.66.1

# YAML parser for config file
pyyaml==6.0.1